pyproject.toml
[tool.poetry]
name = "language-museums"
version = "0.1.0"
description = "Auxiliary code to help with the 'language-museums' project"
authors = ["Joe Germuska <joe@germuska.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
unicode-slugify = "^0.1.3"
ipython = "^7.12.0"

[tool.poetry.dev-dependencies]
ipython = "^7.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
